requirements.txt
Django==4.0.6
Pillow==9.2.0
razorpay==1.3.0
boto3==1.24.42
botocore==1.27.42
django-storages==1.12.3
python-decouple==3.6